+++ R/internal-docs.R
@@ -20,7 +20,7 @@
 #'   map when plotting factor variables. The length of this vector must match
 #'   the number of levels in your factor variable, or num_colors for a
 #'   continuous variable that will be discretized by the function, and the order
-#'   should match the order of the levels of in your factor variable.
+#'   should match the order of the levels in your factor variable.
 #' @param projection One of the following: "cartesian", "mercator", "robinson",
 #'   or "albers", for the equirectangular, Mercator, Robinson, and Albers Equal
 #'   Area projections, respectively. When using the Mercator projection for